Fix keyboard remap and redesign isModifier.
This commit fixes a case where the keyboard doesn't detect the right
modifier when the user remaps a modifier key to a different key,
e.g. CapsLock to Control.
