+++ src/Std/readLine.ts
@@ -1,7 +1,6 @@
 import { Screen } from "../Screen";
 import { Keyboard } from "../Keyboard";
 import { TypeListener, VoidListener } from "../Keyboard/Keyboard";
-import { getIsModifierKey } from "../Keyboard/isModifierKey";
 import { Vector } from "../Toolbox/Vector";
 import { TextBuffer } from "../TextBuffer";
 import { KeyCode } from "../Keyboard/types";
